move test_*tokenizer to /test/tokenizer
