+++ test/CMakeLists.txt
@@ -124,7 +124,7 @@ if(WITH_TESTING)
   add_subdirectory(rpc)
   # add_subdirectory(sequence)
   add_subdirectory(standalone_executor)
-  # add_subdirectory(tokenizer)
+  add_subdirectory(tokenizer)
   # add_subdirectory(white_list)
   add_subdirectory(xpu)
 endif()
